* MDF fix idhash not public
We accidentally made idhash not public by not publishing its header
in the right place.  This is really generic utility stuff, so we
have posted it in the nng/supplemental/util/ directory.

We've also removed the ability to remove this -- its a very small amount
of additional code, as its just a wrapper on top of mandatory functionality.

Signed-off-by: jaylin <jaylin@emqx.io>